% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biomod_data_package_list.R
\name{create_biomod_data_package_list}
\alias{create_biomod_data_package_list}
\title{Create biomod2 data package list}
\usage{
create_biomod_data_package_list(spp_occurence_raster_list, biovar_stack)
}
\arguments{
\item{spp_occurence_raster_list}{List of species occurence rasters}

\item{biovar_stack}{raster::stack of bioclimatic explanationary values}
}
\value{
A list of biomod2 data packages
}
\description{
TODO: set aside data for model evaluation!
}
