% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biomod_ensemble_model_list.R
\name{create_biomod_ensemble_model_list}
\alias{create_biomod_ensemble_model_list}
\title{Create biomod2 ensemble of models}
\usage{
create_biomod_ensemble_model_list(biomod_model_list)
}
\arguments{
\item{biomod_model_list}{List of biomod2 model objects}
}
\value{
A biomod2 ensemble model object
}
\description{
Create biomod2 ensemble of models
}
