% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biomod_ensemble_model_projection_list.R
\name{create_biomod_ensemble_model_projection_list}
\alias{create_biomod_ensemble_model_projection_list}
\title{Project biomod ensemble model onto biovar stack}
\usage{
create_biomod_ensemble_model_projection_list(biomod_ensemble_model_list,
  bimod_model_projection_list)
}
\arguments{
\item{biomod_ensemble_model_list}{List of calibrated biomod2 ensemble models}

\item{bimod_model_projection_list}{List of raster::stacks representing the model projections on biovariables}
}
\description{
Project biomod ensemble model onto biovar stack
}
