% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biomod_model_list.R
\name{create_biomod_model_list}
\alias{create_biomod_model_list}
\title{Create biomod model list}
\usage{
create_biomod_model_list(biomod_data_package_list, biomod_model_options)
}
\arguments{
\item{biomod_data_package_list}{List of biomod2 data package objects}

\item{biomod_model_options}{Biomod2 model options object}
}
\value{
List of biomod2 model objects
}
\description{
Create biomod model list
}
