% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biomod_model_projection_list.R
\name{create_biomod_model_projection_list}
\alias{create_biomod_model_projection_list}
\title{Project biomod2 model onto biovariable stack}
\usage{
create_biomod_model_projection_list(biomod_model_list, biovar_stack,
  proj_name = "historical")
}
\arguments{
\item{biomod_model_list}{List of calibrated biomod2 models}

\item{biovar_stack}{A raster::stack representing the biovariables for the projection}

\item{proj_name}{A charchter name to label the output rasters}
}
\value{
A raster::stack containing projected model representations
}
\description{
Project biomod2 model onto biovariable stack
}
