% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biovar_stack.R
\name{create_biovar_stack}
\alias{create_biovar_stack}
\title{Create masked bioclimatic variable raster stack}
\usage{
create_biovar_stack(region = "biovar_Sweden",
  biovar_list = c("biovar01", "biovar02", "biovar03", "biovar04",
  "biovar05", "biovar06", "biovar07", "biovar08", "biovar09", "biovar10",
  "biovar11", "biovar12", "biovar13", "biovar14", "biovar15", "biovar16"),
  scenario_list = c("rcp45", "rcp85", "rcp26", "rcp60"),
  path = "inst/ext_data/bioclimatic", mask_by_land = T,
  pca_transform = T)
}
\arguments{
\item{region}{String representing the region directory}

\item{biovar_list}{List of string representing the biovar codes}

\item{path}{String representing the path to the bioclimatic directory}
}
\value{
A named list with the region as first level containing lists of;
data file paths, processed historical and future biovariable rasters,
polygon of land, the biovariable and scenario lists used, and optionally
a pca model of the historical biovariables.
}
\description{
Select historic and future  bioclimatic variable rasters from directory,
create ensembles, create a stacks, optionally convert to principle components
and optionally mask by land. Note this function can take a while....
}
\examples{
bvso <- create_biovar_stack(
region='biovar_Sweden',
biovar_list= c("biovar01","biovar02","biovar03","biovar04","biovar05","biovar06",
            "biovar07","biovar08","biovar09","biovar10","biovar11","biovar12",
                        "biovar13","biovar14","biovar15","biovar16"),
                        scenario_list=c("rcp45", "rcp85", "rcp26", "rcp60"),
                        path="inst/ext_data/bioclimatic",
                        mask_by_land = T,
                        pca_transform = T
                        )
                        names(bvso)
                        names(bvso[[1]])
}
