% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ensemble_model_raster_stack.R
\name{create_ensemble_model_raster_stack}
\alias{create_ensemble_model_raster_stack}
\title{Create ensemble model projections raster stack}
\usage{
create_ensemble_model_raster_stack(spp_list, proj_nm)
}
\arguments{
\item{spp_list}{List of species names used in modelling}

\item{proj_nm}{The projection name}
}
\value{
A raster::stack with probalistic (CA, CV), and mean and median binary predictions.
}
\description{
Create ensemble model projections raster stack
}
