% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid_from_poly.R
\name{create_grid_from_poly}
\alias{create_grid_from_poly}
\title{Create ISEA-3-HEXAGON_grid_from_polygon}
\usage{
create_grid_from_poly(polygon_object, grid_res, set_seed = F)
}
\arguments{
\item{polygon_object}{A sf polygon object}

\item{grid_res}{The desired grid resolution for the hex grid}
}
\value{
A sf::sf object where each feature is a hexagon with a unique uuid, area [m2], and perimeter [m]
}
\description{
Create ISEA-3-HEXAGON_grid_from_polygon
}
\examples{
bv <- create_biovar_stack(
region = 'biovar_Sweden',
biovar_list = c("biovar04", "biovar06", "biovar16"),
time_period = "biovar_historical",
path = "inst/ext_data/bioclimatic"
)
g <- create_grid_from_poly(bv$land_poly, grid_res=11, set_seed=999)
if(require(tmap)){
tmap_mode("view")
tm_shape(g)+ tm_fill("area_m2", palette = sf::sf.colors(3), alpha=0.7, colorNA=NULL,
legend.show = T)}else{plot(g)}
}
