% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spp_occurence_raster_list.R
\name{create_spp_occurence_raster_list}
\alias{create_spp_occurence_raster_list}
\title{Create spp presence-absence raster(s)}
\usage{
create_spp_occurence_raster_list(spp_occurence_sfc_list, r, land_mask)
}
\arguments{
\item{r}{Raster object to use as base for adding presence-absence}

\item{land_mask}{sf::sf polygon land mask}

\item{spp_occurence_sf_list}{List of sf::sfc objects with spp occurence}
}
\value{
A list of raster::raster objects representing presence (1) and absence (0) of a species, and NA values for no data.
}
\description{
Create spp presence-absence raster(s)
}
