% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spp_occurence_sf_list.R
\name{create_spp_occurence_sf_list}
\alias{create_spp_occurence_sf_list}
\title{Title}
\usage{
create_spp_occurence_sf_list(spp_list, bbox, creds,
  path = "inst/ext_data/species_gbif", continent = "EUROPE",
  start_year = 1986, end_year = 2020)
}
\arguments{
\item{spp_list}{List of scientific species names}

\item{bbox}{sf::bbox object, used to crop extent of occurences}

\item{creds}{GBIF user credentials}

\item{continent}{The continent to limit search}

\item{start_year}{Start year for occurence query}

\item{end_year}{End year for occurence query}
}
\value{
A named list of sf::sf data.frame(s) per species name.
}
\description{
Title
}
