% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spp_query_list.R
\name{create_spp_query_list}
\alias{create_spp_query_list}
\title{Create list of GBIF queries}
\usage{
create_spp_query_list(spp_list, continent = "EUROPE", creds,
  start_year = 1986, end_year = 2020)
}
\arguments{
\item{spp_list}{Species names}

\item{continent}{The continent
geom WTK polygon bounding box to limit search [THIS DOES NOT SEEM TO WORK]}

\item{creds}{gbif credentials}

\item{start_year}{Start year  to limit search}

\item{end_year}{End year to limit search}
}
\value{
List of rgbif::occ_download_prep objects named by spp_list
}
\description{
Create list of GBIF queries
}
