% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bioclimatic.R
\name{parse_bioclimatic}
\alias{parse_bioclimatic}
\title{Parse directory structure of bioclimatic rasters}
\usage{
parse_bioclimatic(path = "inst/ext_data/bioclimatic")
}
\arguments{
\item{path}{Path to directory with sub-directories of regions each containing historical and future raster files}
}
\value{
A data frame with the directory names, file info and base filename
}
\description{
Parse directory structure of bioclimatic rasters
}
\examples{
df_bc <- parse_bioclimatic(path="inst/ext_data/bioclimatic")
summary(df_bc[[1]])
head(df_bc[[1]])
}
