% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_stats_poly.R
\name{zonal_stats_poly}
\alias{zonal_stats_poly}
\title{Zonal stats for each polygon feature}
\usage{
zonal_stats_poly(fun, polygon_object, raster_object, remove_na = T,
  out_path = F)
}
\arguments{
\item{fun}{Function to apply to values in each polygon}

\item{polygon_object}{A sf polygon feature collection}

\item{raster_object}{A raster::raster object}

\item{remove_na}{Boolean, remove features with NA values?}

\item{out_path}{Write object to this path}
}
\value{
A sf::sf object identical to the input polugon object with zonal statistics added
}
\description{
Zonal stats for each polygon feature
}
\examples{
# add example
}
