% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biovar_stack.R
\name{create_biovar_stack}
\alias{create_biovar_stack}
\title{Create masked bioclimatic variable raster stack}
\usage{
create_biovar_stack(region = "biovar_Sweden",
  biovar_list = c("biovar04", "biovar06", "biovar16"),
  time_period = "biovar_historical",
  path = "inst/ext_data/bioclimatic")
}
\arguments{
\item{region}{String representing the region directory}

\item{biovar_list}{List of string representing the biovar codes}

\item{time_period}{String representing the time-period code}

\item{path}{String representing the path to the bioclimatic directory}
}
\value{
A list with 'biovar_stack' and 'land_poly' representing the raster::stack
of the selected bioclimatic variables, for the region and time-period, masked by land mass
, and a polygon of land, respectively.
}
\description{
Select bioclimatic variable rasters from directory, create a stack and mask by land
}
\examples{
s <- create_biovar_stack(region='biovar_Sweden',
biovar_list= c("biovar04","biovar06","biovar16"),
time_period="biovar_historical",
path="inst/ext_data/bioclimatic")
raster::plot(s$biovar_stack)
}
